+++ lib/src/exui.dart
@@ -1,2 +1,3 @@
 export 'widget/widget.dart';
 export 'string/string.dart';
+export 'icons/icons.dart';