+++ lib/src/icons/icons.dart
@@ -1,0 +1,2 @@
+export 'icon.dart';
+export 'icon_data.dart';